% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisja.R
\name{emisja}
\alias{emisja}
\title{Obliczenia emisji spalin}
\usage{
emisja(
  data = dane_wej,
  kategoria = "Passenger Cars",
  paliwo = "Petrol Hybrid",
  typ = "Small",
  stan = "Euro 4",
  tech = "PFI",
  sub = "CH4"
)
}
\arguments{
\item{kategoria}{character}

\item{dane}{dataframe}

\item{euro}{character}

\item{mode}{character}

\item{substancja}{character}
}
\value{

}
\description{
Obliczenia emisji spalin
}
