% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input}
\alias{input}
\title{Plik do wczytywania danych zewnetrznych potrzebnych do dzialania pakietu}
\format{
Ramka danych zawiera:
}
\source{
Plik Wskazniki zawiera dane emisji z transportu drogowego EEA - Europejska Agencja Srodowiska
}
\usage{
input
}
\description{
Plik do wczytywania danych zewnetrznych potrzebnych do dzialania pakietu
}
\examples{
\dontrun
}
\keyword{datasets}
