% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres.R
\name{wykres}
\alias{wykres}
\title{Wizualizacja danych - wykres}
\usage{
wykres(out, x, y, z)
}
\arguments{
\item{out}{double}

\item{x}{double}

\item{y}{double}

\item{z}{double}
}
\value{
plot
}
\description{
Wizualizacja danych - wykres
}
\examples{

}
