% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisja.R
\name{emisja}
\alias{emisja}
\title{Obliczenia emisji spalin}
\usage{
emisja(
  dane = input,
  kategoria = "Passenger Cars",
  paliwo = (c("Petrol", "Petrol Hybrid", "Diesel")),
  typ = (c("Small", "Mini")),
  stan = "Euro 4",
  tech = (c("PFI", "DPF")),
  sub = "CH4"
)
}
\arguments{
\item{dane}{dataframe - dane wejsciowe}

\item{kategoria}{character - kategoria pojazdu - Passenger Cars,Light Commercial Vehicles,Heavy Duty Trucks,Buses,L-Category}

\item{paliwo}{character - Petrol, Diesel, Petrol Hybrid, LPG Bifuel ~ LPG, LPG Bifuel ~ Petrol, CNG Bifuel ~ CNG, CNG Bifuel ~ Petrol, CNG, Biodiesel}

\item{typ}{character - segment pojazdu - Mini, Small, Medium, Large-SUV-Executive, itp}

\item{stan}{character - norma dopuszczalnych emisji spalin - Euro 1, Euro 2, Euro 3, Euro 4, Euro 5, Euro 6, itp}

\item{tech}{character - technologia ograniczenia spalin - GDI, PFI, GDI+GPF, DPF, DPF+SCR, LNT+DPF, SCR, EGR}

\item{sub}{character - rodzaj emitowanego zanieczyszczenia - CO, NOx, VOC, PM Exhaust, EC, CH4, NH3, N2O}
}
\value{

}
\description{
Obliczenia emisji spalin
}
\details{
Wzor do obliczen emisji:
 Emisja = Nat x ((Alpha  x Procent^2 + Beta x Procent + Gamma + (Delta/Procent)) /
 (Epsilon x Procent^2 + Zita x Procent + Hta) x (1- Reduction))
}
