% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisja.R
\name{emisja}
\alias{emisja}
\title{Obliczenia emisji spalin}
\usage{
emisja(
  dane = input,
  kategoria = "Passenger Cars",
  paliwo = (c("Petrol", "Petrol Hybrid", "Diesel")),
  typ = (c("Small", "Mini")),
  stan = "Euro 4",
  tech = (c("PFI", "DPF")),
  sub = "CH4"
)
}
\arguments{
\item{dane}{dataframe}

\item{kategoria}{character}

\item{paliwo}{character}

\item{typ}{character}

\item{stan}{character}

\item{tech}{character}

\item{sub}{character}
}
\value{

}
\description{
Obliczenia emisji spalin
}
