% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres.R
\name{wykres}
\alias{wykres}
\title{Wizualizacja danych - wykres}
\usage{
wykres(x, dane = out, y = .data$Emisja, z = .data$Nat)
}
\arguments{
\item{x}{double - dowolny wybrany wskaznik przez uzytkownika}

\item{y}{double - obliczona emisja spalin}

\item{z}{double - natezenie ruchu}

\item{out}{double - zestawienie wyników otrzymanych z funkcji 'emisja'}
}
\value{
plot
}
\description{
Wizualizacja danych - wykres
}
